pyproject.toml
[project]
name = "oisol"
dynamic = ["version"]
requires-python = ">= 3.12"
dependencies = [
  "aiohttp",
  "beautifulsoup4",
  "discord.py==2.5.0",
  "cryptography",
  "libretranslatepy",
  "lxml",
  "python-dotenv",
  "requests",
]

[dependency-groups]
qa = [
  "pytest",
  "pytest-xdist",
  "pytest-asyncio",
  "ruff"
]

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"